Separate types out into their own files (#1)
I've split up `src/index.ts` so that each type is in its own file. And
to avoid a giant list of files in one directory, I've grouped the types
a little bit where it makes sense (e.g. all types to do with tables are
in the `src/types/tables` directory). While doing that I've made a some
other tweaks:

- I've dropped the JSF prefix on every exported type. `JSFWorkbook` is
  now `Workbook`, and so on. Use a namespace import (for example,
  `import * as jsf from "jsf"`) if you need to prefix the types

- I renamed a few things:
  - `JSFSheetDefaults` to `WorksheetDefaults`, to be more consistent
    with `Worksheet`
  - `JSFNameDefinition` to `DefinedName`, to be more consistent with
    Excel
  - `JSFExtSheet` to `ExternalWorksheet`, to avoid the abbreviation and
    to match `Worksheet` and `WorksheetDefaults` if we want

- I've allow shadowing of global variables in the ESLint config. Without
  it, ESLint complains about the External and Comment types. This
  matches the config in @borgar/xlsx-convert.