Add GitHub Actions workflow for package releases
Releases a new version when a semver tag is pushed to GitHub. If
successful, it will also create a release on the GitHub repo, and
generate release notes.

This uses the "trusted publishing" process for publishing a package
without relying on long-lived npm tokens.

https://docs.npmjs.com/trusted-publishers
https://github.blog/changelog/2025-09-29-strengthening-npm-security-important-changes-to-authentication-and-token-management/

Trusted publishing requires npm CLI version 11.5.1 or later, which is
why there are minor changes to the lock file.