Add workbook/worksheet views to support edit state
Adds a mandatory `views` property to `Workbook`, and an optional `views`
property to `Worksheet`.

`Workbook.views` is an array that contains one or more `WorkbookView`,
each of which stores display settings for the workbook. Currently, this
is limited to storing only the currently active worksheet. But this
single setting allows JSF to store which of the sheets a user was last
looking at.

`Worksheet.views` is an array that contains zero or more
`WorksheetView`, each of which stores display settings for the
worksheet. Again, this is currently limited to storing the sheet's
currently selected cell and the currently selected range. Both are
optional. This allows JSF to store which cell/range a user has last
selected.

Yes, a workbook and a worksheet can have multiple views. This is covered
in the Office Open XML standard (ISO/IEC 29500:2016) in §18.2.1,
§18.2.30, §18.3.1.88, and §18.3.1.87, which cover the `bookViews`,
`workbookView`, `sheetView`, and `sheetView` elements from the .xlsx
XML. Note that this is a minimal implementation of workbook/worksheet
views, as the spec allows them to store a lot more settings than just
active sheet/cell/range. But it's a start.

What exactly is a view?
=========================

When viewing a workbook (and its worksheets), you can choose various
settings for how it all looks: which sheet is selected, whether it has
gridlines and rulers, what size the cells and cell text are (aka zoom),
and so on.

You can have more than one of these views per workbook. Each view
corresponds to a separate window in the application, and each view has a
different configuration. Yes, Excel lets you have the same sheet open in
different windows at the same time.

https://support.microsoft.com/office/0eea3dc5-d7d1-44c5-a953-25ebfbd6c1a6

How these views are saved in the .xlsx is defined in the Office Open XML
standard. Within a workbook (`xl/workbook.xml` in the zip):

- The root `workbook` element has a mandatory `<bookViews>` child
- The `<bookViews>` has one or more `<workbookView>` children
- A `<workbookView>` has ~10 optional attributes to to describe the view
  configuration

Within a worksheet (e.g. `xl/worksheets/sheet1.xml` in the zip):

- The root `<worksheet>` element has an optional `<sheetViews>` child
- A `<sheetViews>` element has one or more `<sheetView>` children
- A `<sheetView>` element has a required `workbookViewId` attribute,
  plus optional child elements and/or optional attributes that define
  the settings for the worksheet view

Why add views?
================

Supporting workbook/worksheet views allows us to add support for the
user's edit state. While many of the view settings are obscure (does
anyone need to specify the ratio between the tab bar and the horizontal
scroll bar?!), last selected sheet and last selected cell are genuinely
useful. With them, JSF can take a big step towards supporting full Excel
behaviour in a spreadsheet renderer.

Implementation notes
======================

- Because there are now more types concerning workbooks and worksheets,
  I've grouped new and existing types in sub-directories
  (`src/types/workbooks` and `src/types/worksheets`)
- We already have `showGridLines` as a top-level property on the
  `Worksheet` type, but this should actually be stored within the
  worksheet's view. I haven't done anything with that property here, but
  we should consider deprecating it or aliasing it (or something else?)
- Chart sheet views are also a thing (§18.3.1.86), but I haven't
  considered them here
- A workbook has many views, and a worksheet has many views. But each of
  a _worksheet's_ views must be linked to a different _workbook_ view
- Workbook/worksheet views are of course fully supported by Excel.
  LibreOffice Calc will only open the first view of a workbook, and it
  will warn you if you try to save the file. If you do save, you will
  lose all but the first workbook view. Apple Numbers is similarly
  lossy: it only supports one view, and will lose all but the first if
  you export as .xlsx (with no warning)