Adds support for worksheet zoom levels
Builds on top of #4 to add support for including active layout and
layout zoom scales to a worksheet view.

Layouts and zoom scales, you say?
===================================

Excel (and the .xlsx format) support three different layouts for a
worksheet: normal, page layout, and page break preview (sometimes called
sheet layout view). The normal layout is the one everybody's used to;
the other two are most useful when printing a spreadsheet.

The .xlsx format allows applications to store which layout the user last
used, so it can open that layout when the user returns to the sheet.
Each layout also has its own zoom level (10%-400%), and these are all
stored per worksheet view --- Excel and LibreOffice Calc let you have
the same sheet open in different windows at the same time, each with
different display configurations (aka "views"):

https://support.microsoft.com/office/0eea3dc5-d7d1-44c5-a953-25ebfbd6c1a6

These last-used layout and the zoom scales are specified in §18.3.1.87
of the Office Open XML standard (pp. 1700-1701). The OOXML spec calls a
layout a "view" or "view type", but because that term is overloaded (see

Implementation
================

An example of how a worksheet's layouts and zooms might look in an .xlsx
file:

    <worksheet>
      <sheetViews>
        <sheetView
          view="pageLayout"
          zoomScale="80"
          zoomScaleNormal="90"
          zoomScalePageLayoutView="80"
          workbookViewId="0" />
        <sheetView workbookViewId="1" />
      </sheetViews>
    </worksheet>

This gives us a sheet with two views:

1. The active view, with workbookViewId 0, that opens on the page layout
   view with a zoom/scaling factor of 80%. But note that when a user
   switches to normal view, the zoom/scaling factor will be 90%.
2. A second view, workbookViewId 1, where everything is at default
   values (i.e. normal view scaled to 100%)

Using the implementation here, we would represent these XML layouts and
zooms in JSF like this:

    {
      "sheets": [
        {
          "name": "Sheet1",
          "views": [
            {
              "workbookView": 0,
              "activeLayout": "pageLayout",
              "layoutScales": { "normal": 90, "pageLayout": 80 }
            },
            { "workbookView": 1 }
          ]
        }
      ]
    }

Notes
=======

- Excel duplicates the active layout scale value --- it includes it as
  the zoomScale and the zoomScaleXXX attributes. JSF doesn't need the
  active layout separately since it's included via
  WorksheetView.activeLayout
- The spec says zoom scales are restricted to the range 10 to 400. These
  are sensible limits. If a JSF file is being round-tripped is should
  probably clamp the value when it's outside this range
- LibreOffice Calc supports zooms and last-used layout, but has limits
  on multiple sheet views (see #4)
- Apple Numbers ignores any zoom levels  when you open a .xlsx file, and
  doesn't seem to support page layout or page break preview
- Chart sheet views (§18.3.1.86) also use zoomScale, but this isn't
  considered here